delete the db file only if it exists

+++ database/createDb.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 echo "rm existing db"
-rm rpg.db
+[ -f rpg.db ] && rm rpg.db
 
 echo "creating db structure"
 sqlite3 -init structure.sql rpg.db '.quit'
